build: simplify dependencies using `cfg(unix)`
Fixes #33.  Rather than trying to guess at all of the different
concrete target types that are unix, we can lean on rust a bit
more to tell us if our target is a unix platform.

This way, when new targets are added to the ecosystem they
should just work with this crate.

If windows support is added with different dependencies, they can
be added as simply as:

    [target.'cfg(windows)'.dependencies]

Signed-off-by: Paul Osborne <paul.osborne@smartthings.com>